inference/requirements.txt
onnxruntime-gpu
opencv-python
pillow
torchvision
numpy
